+++ examples/java/cdap/salesforce/src/main/java/org/apache/beam/examples/complete/cdap/salesforce/README.md
@@ -15,14 +15,6 @@
     under the License.
 -->
 
-# Apache Beam pipeline examples for CdapIO and CDAP plugins
-
-This directory contains set of [Apache Beam](https://beam.apache.org/) pipeline examples to read data
-from a [CDAP plugin](https://github.com/data-integrations) and write data into .txt file (and vice versa).
-
-Supported CDAP plugins:
-- [Salesforce](https://github.com/data-integrations/salesforce)
-
 ## Gradle preparation
 
 To run this example your `build.gradle` file should contain the following task to execute the pipeline:
@@ -41,7 +33,7 @@ task executeCdap (type:JavaExec) {
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.CdapSalesforceToTxt \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.salesforce.CdapSalesforceToTxt \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
 
@@ -66,7 +58,7 @@ Please see CDAP [Salesforce Batch Source](https://github.com/data-integrations/s
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.TxtToCdapSalesforce \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.salesforce.TxtToCdapSalesforce \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
 
@@ -95,7 +87,7 @@ Please see CDAP [Salesforce Batch Sink](https://github.com/data-integrations/sal
 Gradle 'executeCdap' task allows to run the pipeline via the following command:
 
 ```bash
-gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.CdapSalesforceStreamingToTxt \
+gradle clean executeCdap -DmainClass=org.apache.beam.examples.complete.cdap.salesforce.CdapSalesforceStreamingToTxt \
      -Dexec.args="--<argument>=<value> --<argument>=<value>"
 ```
